Opinionated heading standardization
This move uniformly adopts the conventions summarized in the sphinx docs
(https://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html?highlight=chapter#sections).
While it really is arbitrary, I think there is value in doing it
uniformly.

I also introduce a uniform .../index.rst file for all chapters that can host the
preamble.

There is still an open issue in that sphinx treats files as more than just arbitrary
structures, i.e. whatever headline definition is used, the top-most one
in each file becomes the title (etc.). This is also reflected in the PDF
-- each file represents a "chapter", hence conflicts with the TOC
defined in contensts.rst.inc.

We should standardize this as well, to make it less confusing.

One way would be to really consider each file to be a chapter and treat
the current "chapters" (as defined in the TOC) as "parts" of the book --
which also makes sense to me. If that shall be the way, I'd propose to
frame the introduction a bit differently (to make the difference to
"basics" more pronounced -- pretty much as a guide to the book itself.
It could have "why should I read this?" or "for whom which part this is
written", "what do I need to work with the book", "how can I
contribute". Bits like "philosophy" would then possibly need to be moved
into the "basics" part.

Depending on the feedback to this I would rework this change.