Sketch of a fully custom `gitusernote` directive
This is what is needed to be able to exercise pretty much anything that
we could think of in terms of presentation and behavior of these notes.
At the moment it merely mimics what was there before. What is different
is that all formatting is now done fully manual, so the new code exposes
all relevant interfaces.

In order to achieve this, we now have a Python package for the book
included in the repo.

Now we can start thinking about how to present this content best:
folded, optionally fully removed from a PDF by, configuration...

In the same spirit we can also consider adding additional custom
directives for the book (e.g. for exercises, ...)