Auto-convert figures from SVG to optimal format
Instead of having to include heavy PNGs as a compromise between
HTML and Latex demands.

This requires an additional Python package (requirements.txt adjusted),
and a commandline tool: rsvg-convert

On Debian-like systems this can be installed via:

  sudo apt-get install librsvg2-bin

Readthedocs also supports this.

Sidenote on SVG-compatibility: rsvg-convert (unlike inkscape) does not
support "flowed text". This would express itself by such text boxes not
showing up at all in the converted figures. This can be easily mitigated
by converting such text boxes in Inkscape (Text -> Convert to Text), and
re-saving the SVG.

Usage is demoed in the diff, PNG image is replaced with an editable SVG.