ENH: some wording choices and findoutmore about git log
I have kept myself from changing "write" to "save" when talking about
saving text to a file.  Save is more familiar term to everyone since
that is the  Menu  item they would see in 99% of the editors they
have used so far.  But I do appreciate the thin line here to not
confuse with DataLad save.  But it should not be just asssumed
right away that  "save" would always mean "save into DataLad (history)"
so I think it is good to use more explicit wording here