Recaptured code examples from new execution environment
Changes look sensible (due to env change, but also latest datalad and
git-annex).

There is one open issue: We are using `ls -l` in some places that
reveals thatthe effective user ID creating files on the machine is
not the fake one we configured. We should check whether it is important
to list this information. If so, we may need to configure a dedicated
user.