provide a hidden intro page for basics and usecases
This way, when navigating with the side bar, readers will not 'accidentally' find themselves in
the Basics or the Usecases, but stumble across an index page that makes it explicit that the
previous book part ended.

The index pages do not show up in the general toctree.

We could potentially list use cases in the use case index file with a short summary, for a bit
of navigation help.

WIP

WIP