PDF: make frontpage reflinks to sections in the intro work
Links to html pages are not resolved, but are nice for custom
descriptions (i.e., not the section heading). This adds
separate paragraphs with links for PDF and HTML output.
It also moves the user types out of the intro part (where it
was hidden in HTML, but right at the start for PDF) into
the Appendix