BF: Replace inaccurate description of push behavior
As described in #481, the previous description of how push
behaves in datasets with multiple branches was inaccurate,
and resulted from my explorations of push with an explicitly
configured dataset. This change rectifies this, and instead
shows how to configure datasets such that more than just the
current branch is pushed. Fixes #481.