Make: rules for removing stand-alone DVC section
This adds a separate clean rule for non-narrative sections.
At the moment, it only contains DVC. This section does not have any
relation to the DataLad-101 narrative, and hence doesn't need
to be redone with a complete rebuild of the book.
Other such unrelated sections can be added to this rule (which could also
be renamed), but so far, this section is the only one in the book
(apart from usecases) that does not depend on the narrative.

Rebuilding DVC is also tedious, because this section also pushes back to
GitHub-based repos