No vertical lines in a table
This is a first demo of how it could be done. Essentially given
sphinx a dedicated clue via `tabularcolumns` instead of having it
guess.

https://www.sphinx-doc.org/en/master/usage/restructuredtext/directives.html#directive-tabularcolumns

It is using the sphinx-specifc `\Y` column width specifier that should
work across all table type we need (sphinx is switching type a lot,
depending on table content).