Complete workflow to rebuild runrecords
In less than 1k log lines, this makes a (full) rebuild of the handbook,
including execution of runrecords.

It uses the standard build setup documented for the handbook.

Any resulting diff to the `doc/` directory (i.e. updated runrecords) is
visualized at the end of the log, and is made available for download
at a URL that is also shown at the very end.

With this diff contributors/maintainers can decide to incorporate the
changes, in full or partially into their PRs.