Make storage folder name consistent
Previous code block creates a directory called `dvc-storage`, so the
current block should use it too (hyphen is also consistent with
`datalad-sibling`). As a side note, the folder containing the dvc
remote does not need to exist in advance - when running locally and
"by hand", the command created the `dvc_remote` folder for me.

This change will require an update to run records.