Make branch and tag names consistent and unambiguous
This changes the DVC workflow to use sgd-pipeline & random-forest as
branch names, and sgd & randomforest as tag names.

This change improves consistency between branches, but also removes
the mixing of hyphens and underscores between code and
narrative. Additionally, it removes the chance to run into git
warnings when switching branches (although it's perfectly fine to use
the same name for tag and branch, git will warn about ambiguous refs).