Migrate fix from work on print version
These have accumulated over time and are now moved in bulk. This is a
hand-edited diff that aims to only pick the changes that are also
beneficial for the online version. These include:

- Shorten URLs (remove trailing slashes, avoid needless deep-links)
- Remove unnecessarily wordy passages
- Updated listings
- Transition from `master` to `main` branch
- Attempt to standardize term capitalization
- Remove aged advice on transitioning from old versions

Plus various additions that add new info or pointers to more recent
developments.

Givent he size of the diff and its semi-manual creation, it should be
noted that the following changes would be unintentional oversights:

- Removing manpage references
- Removing references/links to parts fo the online handbook