New roles `:dlcmd:`, `:gitcmd:`, `:gitannexcmd:`, and `:shcmd:`
These should be used instead of ':command:`datalad ...`' etc. It will
insert a link to the online command manpage on first use for a specific
command in any source document (except for `shcmd`).

Purpose:

- `dlcmd`: any datalad subcommand
- `gitcmd`: any git subcommand
- `gitannexcmd`: any git-annex subcommand
- `shcmd`: any (UNIX-like) command

This makes the manual handling of that insertion (which is presently
different between online and print version) obsolete.

For now, the manpage links for `dlcmd` are only valid for datalad-core
commands (also true for the established `:manpage:` approach).
Additional support for specific extension can be added in code in the
future.