Update DVC workflow for DVC v3
DVC 3.0 was released with a set of breaking changes. However, the only
one which affects us is the removal of `dvc run` in favour of `dvc
stage add --run`. This commit updates the DVC workflow to use the 3.0
command, removing the cause of build failures reported in #965.

This is effectively a reversal of 47b3ed9 which pinned dvc < 3.0 for
our build. While there were multiple quick-release 3.x versions (going
from 3.0 to 3.12 in the course of 3 weeks), they seem to adhere to
semver, as the changelog reports no breaking changes. At a quick
glance, none of the new features or bug fixes are particularly
relevant for our (rather basic) workflow, so it seems ok to go with an
updated major version dependency.