Turn the "eternal" handbook date into symbols
We fix the committer date in the handbook for reasons of reproducibility
(no random shasum changes, simply because we execute code at a different
time). Hence the dates/times shown do not really make sense (they appear
constant, even across multiple steps of a procedure. This appears odd.

This changeset replace the timestamps with symbolic placeholders.

Their meaning still needs to be described in the intro.

Ping #1009