Robustify pip-based installation instructions
Going via `python -m pip ...` is better able to function as expected
in a variety of environments, and does not rely on match PATH
resolution.

This also (largely) simplifies the `pip3` naming complexity on Mac.

Now there could be systems with `python` is named `python3` and that is
all about it.