workaround: adjust checked-out branch to prevent error
For yet to be determined reasons, the DVC workflow has trouble with the master-main
transition. Despite configuring the default branch to main, DVC repos keep master
as their default. The internal bare repositories established as a push target,
however, points to main with its HEAD, resulting in a failure to checkout upon
cloning.
This fix is rather lazy - instead of dealing with the underlying branch confusion
and making DVC repos use main, we explicitly ignore the HEAD from the cloned bare
repository by cloning the master branch specifically from the push target.
This is done in an internal call, and hidden from the reader