Make macOS spelling consistent
It looks like Apple favours the macOS spelling, and according to
https://en.wikipedia.org/wiki/MacOS, "in 2016, with the release of
macOS 10.12 Sierra, the name was changed from OS X to macOS to align
it with the branding of Apple's other primary operating systems".

Note for grep users: "MacOSX" in installation is left
intentionally (part of miniconda installer naming pattern).