Move windows-wit box content into dedicated file
The main goal is to trim the diff to the print version again, where
content was moved and taken out of containers. This makes merged
conflict-ladden and painful. With this approach, any content change
would happen in the dedicated file, which can be used for boxes
and sections alike. Associated index items must stay outside,
because they are declared on the container.

This changeset is complete, the rest will happen in the print-version
branch.