Adjust copybutton setup to ignore '<prompt> #...' lines
This is needed, because for the 'console' lexer to work properly,
comment lines must still follow a prompt. This is a bit "strange",
but actually what it would look like in a console session (ie.,
not a script), and is also needed to distinguish comments from
command output that looks like a comment.