Avoid declaring an `output/` directory as mandatory
It is only important that outputs are not placed into the `inputs/`
directory.

Using an `output/` directory is OK, though. However, it can become
awkward, if such a data module is used as input to another. File paths,
seen from the YODA root would then be: `inputs/outputs/something.dat`

This is a non-invasive change that accounts enough for this aspect,
IMHO.